EcoSentry/requirements.txt
flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.6.0
pillow==8.3.1
matplotlib==3.4.2
seaborn==0.11.1
requests==2.26.0
python-dotenv==0.19.0
joblib==1.0.1
xgboost==1.4.2
geojson==2.5.0
flask-wtf==0.15.1
gunicorn==20.1.0
